requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
langchain_groq

openai
tiktoken
langchain_huggingface
sentence_transformers

pinecone
pinecone_client

pypdf
langchain-text-splitters

ragas
numpy
scikit-learn

